Pi3-main/requirements.txt
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
pillow
opencv-python
plyfile
huggingface_hub
safetensors